requirements.txt
python_dotenv
langchain_groq
Flask
gunicorn
markdown
pygments